ai-service/requirements.txt
flask==2.3.3
langchain==0.1.5
ollama==0.1.2
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
web3==6.11.1
langchain-community>=0.0.17,<0.1
